# api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Synthetic Ascension EHR Platform API
  description: |
    Advanced AI-powered Electronic Health Record (EHR) synthesis platform with comprehensive agentic architecture.
    
    ## Features
    - 50+ specialized agents across 6 categories
    - 6-phase pipeline execution with real-time monitoring  
    - Enterprise-grade privacy and security safeguards
    - Full Langflow integration for visual workflow editing
    - Clinical realism certification and compliance validation
    
    ## Agent Architecture
    - **Cohort Constructor (11 agents)** - Demographics and population modeling
    - **Clinical Journey Generator (11 agents)** - Healthcare pathways and encounters  
    - **Data Robustness & Noise (10 agents)** - Privacy protection and data variability
    - **QA & Validation (13 agents)** - Quality assurance and compliance verification
    - **Explanation & Provenance (13 agents)** - Reporting and audit trail management
    - **Supervision & Orchestration (9 agents)** - System monitoring and coordination
    
    ## Security & Privacy
    - Differential privacy with k-anonymity scoring
    - Clinical realism certification with human-in-the-loop validation
    - HIPAA, FDA, and healthcare regulatory compliance
    - Re-identification risk assessment and mitigation
  version: 3.0.0
  contact:
    name: Synthetic Ascension API Support
    url: https://synthetic-ascension.com/support
    email: api-support@synthetic-ascension.com
  license:
...
servers:
...
paths:
...
  /api/ux/dashboard/recent-jobs:
    get:
      tags: [UX Dashboard]
      summary: Get Recent Jobs
      description: Recent jobs for dashboard display
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 50
          description: Number of recent jobs to return
      responses:
        '200':
          description: Recent jobs retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobCard'
...
components:
  schemas:
...
    JobCard:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        job_name:
          type: string
        status:
          type: string
          enum: [running, completed, failed]
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 100
        current_phase:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
          nullable: true
        population_size:
          type: integer
        condition:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ApiKeyAuth: []
  - BearerAuth: []